Extract runtime router into agent/runtime/router.rs
